--- a/nnvm/src/top/nn/convolution.cc
+++ b/nnvm/src/top/nn/convolution.cc
@@ -414,6 +414,83 @@ NNVM_REGISTER_OP(_contrib_conv2d_winograd_without_weight_transform)
 
 DMLC_REGISTER_PARAMETER(WinogradConv2DParam);
 
+NNVM_REGISTER_OP(_contrib_conv2d_NCHWc_winograd_weight_transform)
+    .describe(
+        R"code(Weight transformation of winograd fast convolution algorithm.
+Separate this into another nnvm symbol in order to enable Precompute Pass to compute the
+weight transformation in advance.
+
+- **weight**: (Packed weight matrix)
+)code" NNVM_ADD_FILELINE)
+    .add_argument("weight", "6D Tensor", "Packed weight tensor.")
+    .add_arguments(WinogradWeightTransformParam::__FIELDS__())
+    .set_attr_parser(ParamParser<WinogradWeightTransformParam>)
+    .set_attr<FGetAttrDict>("FGetAttrDict",
+                            ParamGetAttrDict<WinogradWeightTransformParam>)
+    .set_attr<FInferShape>(
+        "FInferShape",
+        [](const nnvm::NodeAttrs &attrs, std::vector<TShape> *in_shape,
+           std::vector<TShape> *out_shape) {
+          const auto &param =
+              nnvm::get<WinogradWeightTransformParam>(attrs.parsed);
+          const TShape &wshape = (*in_shape)[0];
+
+          CHECK_EQ(wshape.ndim(), 6)
+              << "Packed Weight should be a 6 dimensional tensor";
+
+          // Input kernel layout is essentially COO, CII, KH, KW, CIII, COOO
+          // Transformed kernel layout is COO, CII, CIII, KH, KW, COOO
+          TShape oshape({wshape[0], wshape[1], wshape[4],
+                         param.tile_size + wshape[2] - 1,
+                         param.tile_size + wshape[3] - 1, wshape[5]});
+          NNVM_ASSIGN_OUTPUT_SHAPE(attrs, *out_shape, 0, oshape);
+          return true;
+        })
+    .set_attr<FCorrectLayout>("FCorrectLayout",
+                              [](const NodeAttrs &attrs,
+                                 std::vector<Layout> *ilayouts,
+                                 const std::vector<Layout> *last_ilayouts,
+                                 std::vector<Layout> *olayouts) {
+                                const auto &param =
+                                    nnvm::get<WinogradWeightTransformParam>(
+                                        attrs.parsed);
+                                Layout kernel_layout(param.kernel_layout);
+                                NNVM_ASSIGN_LAYOUT(*ilayouts, 0, kernel_layout);
+                                NNVM_ASSIGN_LAYOUT(*olayouts, 0, kernel_layout);
+                                return true;
+                              })
+    .set_attr<FInferType>("FInferType", ElemwiseType<1, 1>)
+    .set_num_outputs(1)
+    .set_num_inputs(1)
+    .set_support_level(5);
+
+NNVM_REGISTER_OP(_contrib_conv2d_NCHWc_winograd_without_weight_transform)
+    .describe(R"code(Compute conv2d with winograd algorithm.
+
+- **data**: Input is 5 array of shape  (batch_size, in_channel_outer, height, width, in_channel_inner)
+- **weight**: Any shape
+            We do not check shape for this input tensor.
+
+- **bias**: (channels,)
+- **out**:  Output is 4D array of shape (batch_size, channels, out_height, out_width)
+)code" NNVM_ADD_FILELINE)
+    .add_argument("data", "5D Tensor", "Input data.")
+    .add_argument("weight", "6D Tensor", "Transformed weight tensor.")
+    .add_argument("bias", "1D Tensor", "Bias parameter.")
+    .add_arguments(WinogradConv2DParam::__FIELDS__())
+    .set_attr_parser(ParamParser<WinogradConv2DParam>)
+    .set_attr<FGetAttrDict>("FGetAttrDict",
+                            ParamGetAttrDict<WinogradConv2DParam>)
+    .set_attr<FListInputNames>("FListInputNames",
+                               UseBiasListInputNames<WinogradConv2DParam>)
+    .set_attr<FInferShape>("FInferShape", WinogradConv2DInferShape)
+    .set_attr<FInferType>("FInferType", Conv2DInferType<WinogradConv2DParam>)
+    .set_attr<FCorrectLayout>("FCorrectLayout",
+                              Conv2DCorrectLayout<WinogradConv2DParam>)
+    .set_num_outputs(1)
+    .set_num_inputs(UseBiasNumInputs<WinogradConv2DParam>)
+    .set_support_level(5);
+
 NNVM_REGISTER_OP(_conv2d_grad)
   .describe(R"code(2D convolution grad.
 
@@ -440,7 +517,6 @@ NNVM_REGISTER_OP(_conv2d_grad)
 })
 .set_attr<FInferType>("FInferType", ElemwiseType<3, -1>)
 .set_attr<TIsBackward>("TIsBackward", true);
-
 
 DMLC_REGISTER_PARAMETER(Conv2DTransposeParam);