… topi/python/topi/x86/conv2d.py
… topi/tests/python/test_topi_conv2d_NCHWc_winograd.py