This is the implementation alluded to in
https://discuss.tvm.ai/t/improved-direct-winograd-nchwc-cpu-implementation-with-resnet-50-results/

It is a pretty standard Winograd implementation, modified for NCHWc
layout. It achieves reasonable speedups (up to 2x vs current
implementation) on a number of ResNet 3x3 layers on SKL and AVX.

Benchmarks:

- Single-threaded, GCE Skylake instance, ~120GFLOP/s peak fp32 throughput.
- 1000 iterations of AutoTVM tuning per task.

| Workload  | Baseline (GFLOP/s) | Winograd (eff GFLOP/s)
| ------------- | ------------- | ------------ |
| 3x3s1p1, 1x64x56x56 -> 1x64x56x56 | 104 | 193 |
| 3x3s1p1, 1x128x28x28 -> 1x128x28x28 | 110  | 221  |
| 3x3s1p1, 1x256x14x14 -> 1x256x14x14 | 107 | 182 |
| 3x3s1p1, 1x512x7x7 -> 1x512x7x7 | 93 | 176 |

TODO:

- [x] Unit tests.
- [x] Modifications to `tune_nnvm_x86.py`.
- [x] Implement parallelism.
- [x] More benchmarking results.