@masahi 2018-11-15
@ajtulloch you don't need another ```_contrib_conv2d_NCHWc_winograd_without_weight_transform``` symbol. You can do weight transform any way you like inside alter_op_layout. See how ARM winograd does this [here](https://github.com/dmlc/tvm/blob/master/topi/python/topi/arm_cpu/conv2d.py#L569-L572).