@ajtulloch 2018-11-15
@masahi how can I avoid adding `_contrib_conv2d_NCHWc_winograd_without_weight_transform`? This is an operator that takes a 5D NCHWc `input` tensor and a pre-transformed 6D `transformed_weight` tensor, and does a 3x3 Winograd convolution. I don't see how to express that with either `_contrib_conv2d_NCHWc` (need to modify decl to use Winograd method) or `_contrib_conv2d_winograd_without_weight_transform` (doesn't accept NCHWc input tensor).

I agree that we could probably remove the `_contrib_conv2d_NCHWc_winograd_weight_transform` by expressing it as a composition of `conv2d_winograd_weight_transform` + `reshape` + `transpose`.

But just as NCHW Winograd convolution required new NNVM ops, and NCHWc convolution required new NNVM ops, we need new ops for NCHWc Winograd AFAICT.