@masahi 2018-11-15
Have you tried ```_contrib_conv2d_winograd_without_weight_transform``` with NCHWc inputs? I have a home-grown NCHWc winograd + NNVM integration internally.  ```_contrib_conv2d_winograd_without_weight_transform```  worked for me. It does accept 5d inputs.

 As long as you have a correct 
```
    new_attrs['layout'] = 'NCHW%dc' % ic_bn
    new_attrs['out_layout'] = 'NCHW%dc' % oc_bn
```
setup in _alter_conv2d_layout, I think it should work.

For the filter shape, the current solution in nnvm is not to do weight shape check during infer shape (see [here](https://github.com/dmlc/tvm/blob/master/nnvm/src/top/nn/convolution.cc#L180)). So you can use filter transform with any layout.