@ajtulloch 2018-11-15
@masahi doesn't that logic apply to having duplicated conv2d/conv2d_NCHWc ops in NNVM as well? I don't see why it's reasonable to have `conv2d`/`conv2d_NCHWc`/`conv2d_winograd_without_weight_transform` but not also `conv2d_NCHWc_winograd_without_weight_transform`.