@masahi 2018-11-15
I see, honestly I don't know why a new ```sym.contrib.conv2d_NCHWc``` symbol was introduced, even though it uses the same param / infer_shape/ correct_layout as ```sym.conv2d```. @yzhliu Can you comment?

My opinion is that since we already have everything on nnvm side to enable NCHWc winograd, we don't need another duplicated symbol.