@ajtulloch 2018-11-16
@masahi I'd tend to agree, but I can also understand the appeal of being able to say 'conv2d and conv2d_winograd ops take a 4D NCHW input and produce 5D NCHWc output, conv2d_NCHWc and conv2d_NCHWc_winograd ops take a 5D NCHWc input and produce 5D NCHWc output', instead of essentially 'conv2d nodes take any dimension input and produce any dimension output'.