@masahi 2018-11-16
FYI, the recently added cuda int8 convolution uses NCHW4c layout but ```sym.conv2d``` symbol (see [here](https://github.com/dmlc/tvm/blob/master/topi/python/topi/cuda/conv2d_winograd.py#L371-L389)). And as I mentioned, the conv2d_winograd can take a filter transform of arbitrary layout by not doing weight shape check.   

So things are not consistent already. I think it's good time to discuss this issue as we transition to Relay @tqchen  @merrymercy