@ajtulloch 2018-11-16
@yidawang the parallelism part isn't well optimized (we need to conditionally enable parallelism  depending on what granularity we compute the previous stages at), but I think everything else should be good to go. Note the caveats mentioned in https://discuss.tvm.ai/t/improved-direct-winograd-nchwc-cpu-implementation-with-resnet-50-results/1017/16?u=ajtulloch BTW, I don't have good heuristics for avoiding these cases right now so maybe just brute-force disabling it for CIn x COut >= 256 * 256 or something would be reasonable.