@yidawang 2018-11-16
@ajtulloch Perhaps [this paper](https://dl.acm.org/citation.cfm?id=3178496) may have some insight? Optimizing N-Dimensional, Winograd-Based Convolution for Manycore CPUs.