@tqchen 2018-11-20
@ajtulloch is this PR ready for review? please request reviewers