@merrymercy 2018-11-25
The todo item `Modifications to tune_nnvm_x86.py.` is not finished. Please add some examples to the tutorial.