@yzhliu 2018-12-10
@masahi regarding the `sym.contrib.conv2d_NCHWc`, only that the schedule and compute are slightly different. I agree we can merge it into regular conv impl.