@tqchen 2019-01-31
ping @ajtulloch can we land this PR for the 0.5?