@FrozenGene 2019-04-24
@ajtulloch Could you update this work and upgrade it in Relay? I also find it have better performance improvement on ARM compared than default winograd schedule. I suggest we also add it on ARM CPU, for example, named it 'winograd_nchwc' and so on.