@tqchen 2019-07-23
superceded by #3553  thanks to @cbalint13