@cbalint13 2019-07-23
@ajtulloch , @tqchen ,

#3553 is only a part of the integration, will continue with @ajtulloch  NCHWc rebase .